OneDrive/Desktop/udemy-langchain/english-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
notion-client==2.2.1
pydantic==2.5.3
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
